example2/manual/service2/requirements.txt
Flask==3.0.0
opentelemetry-api>=1.20.0,<2.0
opentelemetry-sdk>=1.20.0,<2.0
opentelemetry-exporter-jaeger-thrift>=1.20.0,<2.0
